requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
shap>=0.42.0
umap-learn>=0.5.0
plotly>=5.15.0
jupyter>=1.0.0
ipykernel>=6.25.0
tqdm>=4.65.0
pyyaml>=6.0